Install and run NTPd on all hosts in the cluster.
Try to keep hosts agreeing on time.

Hosts in the cluster will synchronize with hosts in Ansible group `ntp_server`,
and hosts in `ntp_server` will synchronize with 4 randomly-chosen public NTP
server in the `pool.ntp.org` DNS domain.  If no host is in the `ntp_server`
group, then all hosts in the cluster will sync time with public NTP servers.

No matter what the setting of `ntp_server` is, all hosts in the cluster will
pair with other hosts in the cluster.  (Which should prevent them from
diverging, even if time gets out of sync with an external source.)

Fixes #380