Re-work Ceph role according to the "Manual Deployment" section of the Ceph docs.
For reference, see: http://docs.ceph.com/docs/master/install/manual-deployment/

At the moment, installation succeeds on Ubuntu 14.04 and 16.04 with Ceph release
selectable from "hammer" up to (and including) "luminous".  Only the "rbd" pool
will be created; a Ceph MDS is installed and started (if a node with the
`ceph_mds` role is available) but no filesystem is created.