SLURM: Allow configuring more parameters in `slurm.conf` through setup variables.
Specifically, it is now possible to set variables in the `[setup/*]`
section to assign values to the following SLURM configuration
parameters:

* `FastSchedule` (default 1)
* `JobAcctGatherFrequency` (default 60)
* `JobAcctGatherType` (default `jobacct_gather/linux`)
* `MaxArraySize` (default 1000)
* `MaxJobCount` (default 10000)
* `ProcTrackType` (default `proctrack/linuxproc`)
* `ReturnToService` (default 1)
* `SelectType` (default `select/cons_res`)
* `SelectTypeParameters` (default `CR_Core_Memory`)
* `TaskPlugin` (default `task/none`)

The ElastiCluster setup variable name corresponding to a SLURM
parameter name is the lowercased name prefixed with `slurm_`.  For
instance, SLURM parameter `FastSchedule` can be configured via the
variable `slurm_fastschedule`.  (Note that SLURM parameter names are
not case-sensitive, but ElastiCluster variable names are.)

Default values have not changed from previous ElastiCluster releases.