SLURM: Support use of cgroups (opt-in).
Configure cgroup support in SLURM if any one of the cgroup-based
plugins (`task/cgroup`, `jobacct_gather/cgroup`, or
`proctrack/cgroup`) is configured in `slurm.conf`.

*Note:* SLURM's cgroup support requires that swap accounting be
enabled in the kernel.  This is *not* the default on Debian and
Ubuntu, and a reboot is needed to enable it.  ElastiCluster will by
default try to configure the bootloader but *not* reboot the nodes --
use `global_var_allow_reboot=yes` to change this default and reboot
the nodes if needed.