SLURM: Default for `ReturnToService` is now `2`.
If CUDA or the `task/cgroup` plugins are used, it is possible that a
reboot happens during the installation.  When the nodes come back up,
SLURM will mark them as "down" since the reboot was unexpected, and
wait for a sysadmin to issue `sudo scontrol update nodename=... state=resume`.
With `ReturnToService=2`, nodes where `slurmd` is running will be
automatically returned to "idle" state, which is what most
people (likely) want.