AWS: Calculate SSH key fingerprints the same way AWS does
Add support for calculating AWS key fingerprints, which eliminates the
warning on cluster start.  More details on these forum posts:

* https://forums.aws.amazon.com/message.jspa?messageID=307410
* http://stackoverflow.com/questions/21709817/printing-pem-key-der-form-in-python

*Note:* Credits for this code and the above explanation go to @jwm who
originally submitted it in [PR 132][a]; however, by the time I started
to review/merge it, ElastiCluster's code had changed so much that it
was easier to just merge the changes "by hand".

[a]: https://github.com/gc3-uzh-ch/elasticluster/pull/132/files