GCP: Hard-code image name short-cuts, give error if name not well-known.
Previous code would just map unknown image name prefix `X` to a cloud
project name `X-cloud`.  Given that the default names of snapshots are
now something like `image-42`, this has a lot of potential of being
wrong.  So hard-code the list of well-known images and projects, and
error out otherwise.