Raise an error if not enough nodes of each kind are available at setup.
This reverses the previous behavior which was to consider an empty
inventory file as "fine" (!?!?), and, consequently, print out the
message "Your cluster is ready" which was totally misleading.

In addition, remove code that tried to juggle nodes around if not
enough nodes of each kind were started. The reason for this is
threefold:

(1) cloud providers have gotten so much better since 2012 that VM boot
    errors are rare,

(2) this "juggling" behavior was anyway incorrect since different node
    kinds can in principle have different flavors, and

(3) trying to be too smart makes it difficult to reason about what a
    supposedly simple function like `Cluster.start()` does.