Ensure additional file names are passed to Ansible as absolute paths.
This is necessary since the working directory is changed to a
temporary one.  Since we do not know what is a file name and what is
not (without reimplementing all the command-line parsing that
`ansible-playbook` does), we adopt a rather crude heuristics: what
looks like an existing file name (in the invocation directory) is a
file name and is prepended the current working directory (if not
already there).