Adds the ability to pause and unpause instances in the cluster.
This is a good first step, but this still leaves the cluster in
an inconsistent state when restarted - setup needs to be called
externally.  I think that we should probably either fold the
setup call in here, or else add something to the ansible playbooks
that make them re-consistentize the state after a restart.