Migrate links from old web site to new pages.
Many thanks to Maiken Pedersen for noticing the issue and reporting it
on the mailing list!

Some of the changes are in the Beamer/LaTeX source files for old
presentations, which won't -of course- change.  But it still makes
sense to edit them, in case the source files are ever re-used or
remixed into new presentation material.