SLURM: Create account for each non-system user.
Define a non-system user as one that has home directory under `/home`.  This
is very crude but avoids probing for UIDs, which has a different set of
issues:

- the first non-system UID differs across CentOS/RHEL and Debian/Ubuntu
- Some system users like `nobody` often end up having a high UID anyway (e.g., 65534)

The change was actually prompted by the discovery that GCP creates users
corresponding to the SSH keys registered at a project level (menu *Compute
Engine > Metadata > SSH keys*), and the user name comes from the comment field
in the SSH key; therefore, predicting actual user names is impossible and
ElastiCluster should either have ad-hoc code for this case, or attempt to just
make SLURM accounts for anything sensible (which is what is attempted here).