Use GitHub actions for CI (#689)
Replace use of Travis CI with GitHub Actions.  Actions are more flexible and have the advantage that they are copied with the repository, so we don't need to set them up again each time the repository moves or is cloned.

This started out as a straightforward port of the Travis CI configuration, but I ended up with a slightly different organization.  There are 2 main workflows:

- the "Tests" workflow in `tests.yml` which runs:

  * unit tests, for all versions of Python supported by GitHub actions
  * installation tests on all the runner platforms supported by GitHub, using the default `python` binary there

- the "Publish" workflow in `publish.yml` which creates a Docker image; publishing is triggered only on completion of the "Tests" workflow on the `master` branch, and is aborted immediately if any of the tests is not successful.  Pull requests should *not* be published automatically, as contributors may then publish images using secrets stored on the `master` branch.